move monomodrules and custom attributes into subfolder
