+++ Assembly-CSharp/Patches/Attributes/MonoModRules.cs
